Title: 원 모양으로 되어있는 방, Time: 16ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 240304/원 모양으로 되어있는 방/a-room-in-a-circle.cpp
@@ -16,12 +16,12 @@ int main() {
 
     // 시작점을 각 방으로 설정해 놓고 모든 거리 구해보기. 
 
-    long min = INT_MAX;
-    long total = 0;
+    unsigned int min = INT_MAX;
+    unsigned int total = 0;
 
     for (int i = 0; i < n; i++) {
         // 거리 구하는 공식.
-        long distance = 1;
+        unsigned int distance = 1;
         total = 0;
 
         for (int j = i + 1; j < n; j++) {
